Hill Cipher
Hill cipher is a type of substitution cipher. In this type of cipher
encoding process is actually done by dividing the plaintext into units
and replacing those plaintext blocks with cipher text by using linear
algebra.
While decoding even the cipher text is divided into blocks and the
conversion of cipher text blocks into plaintext is done by using same
linear algebra. In this technique, each letter used in the process is
mainly represented by using the number modulo 26. The matrix is used as
a key for the actual encryption and decryption process.